projects/agrostima-scout/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
anthropic>=0.40.0
apscheduler==3.11.2
python-dotenv==1.0.1
